+++ GHEtool/Methods/__init__.py
@@ -1,1 +1,2 @@
-from .optimise_load_profile import optimise_load_profile_power, optimise_load_profile_energy
+from .optimise_load_profile import optimise_load_profile_power, optimise_load_profile_energy, \
+    optimise_load_profile_balance