+++ GHEtool/Methods/optimise_load_profile.py
@@ -344,3 +344,176 @@ def optimise_load_profile_energy(
                    building_load_copy.hourly_cooling_load_simulation_period - borefield_load.hourly_cooling_load_simulation_period))
 
     return borefield_load, external_load
+
+
+def optimise_load_profile_balance(
+        borefield,
+        building_load: Union[HourlyBuildingLoad, HourlyBuildingLoadMultiYear],
+        temperature_threshold: float = 0.05,
+        use_hourly_resolution: bool = True,
+        max_peak_heating: float = None,
+        max_peak_cooling: float = None,
+        dhw_preferential: bool = None
+) -> tuple[HourlyBuildingLoad, HourlyBuildingLoad]:
+    """
+    This function optimises the load for maximum power in extraction and injection based on the given borefield and
+    the given hourly building load, by maintaining a zero imbalance. It does so based on a load-duration curve.
+
+    Parameters
+    ----------
+    borefield : Borefield
+        Borefield object
+    building_load : HourlyBuildingLoad | HourlyBuildingLoadMultiYear
+        Load data used for the optimisation.
+    temperature_threshold : float
+        The maximum allowed temperature difference between the maximum and minimum fluid temperatures and their
+        respective limits. The lower this threshold, the longer the convergence will take.
+    use_hourly_resolution : bool
+        If use_hourly_resolution is used, the hourly data will be used for this optimisation. This can take some
+        more time than using the monthly resolution, but it will give more accurate results.
+    max_peak_heating : float
+        The maximum peak power for the heating (building side) [kW]
+    max_peak_cooling : float
+        The maximum peak power for the cooling (building side) [kW]
+    dhw_preferential : bool
+        True if heating should first be reduced only after which the dhw share is reduced.
+        If it is None, then the dhw profile is not optimised and kept constant.
+
+    Returns
+    -------
+    tuple [HourlyBuildingLoad, HourlyBuildingLoad]
+        borefield load, external load
+
+    Raises
+    ------
+    ValueError
+        ValueError if no correct load data is given or the threshold is negative
+    """
+    # copy borefield
+    borefield = copy.deepcopy(borefield)
+
+    # check if hourly load is given
+    if not isinstance(building_load, (HourlyBuildingLoad, HourlyBuildingLoadMultiYear)):
+        raise ValueError("The building load should be of the class HourlyBuildingLoad or HourlyBuildingLoadMultiYear!")
+
+    # check if threshold is positive
+    if temperature_threshold < 0:
+        raise ValueError(f"The temperature threshold is {temperature_threshold}, but it cannot be below 0!")
+
+    # since the depth does not change, the Rb* value is constant
+    borefield.Rb = borefield.borehole.get_Rb(borefield.H, borefield.D, borefield.r_b,
+                                             borefield.ground_data.k_s(borefield.depth, borefield.D))
+
+    # set load
+    borefield.load = copy.deepcopy(building_load)
+
+    # set initial peak loads
+    init_peak_heating: float = borefield.load.max_peak_heating
+    init_peak_dhw: float = borefield.load.max_peak_dhw
+    init_peak_cooling: float = borefield.load.max_peak_cooling
+
+    # correct for max peak powers
+    if max_peak_heating is not None:
+        init_peak_heating = min(init_peak_heating, max_peak_heating)
+    if max_peak_cooling is not None:
+        init_peak_cooling = min(init_peak_cooling, max_peak_cooling)
+
+    # peak loads for iteration
+    peak_heat_load: float = init_peak_heating
+    peak_dhw_load: float = init_peak_dhw
+    peak_cool_load: float = init_peak_cooling
+
+    # set iteration criteria
+    cool_ok, heat_ok = False, False
+    while not cool_ok or not heat_ok:
+        # limit the primary geothermal extraction and injection load to peak_heat_load and peak_cool_load
+        borefield.load.set_hourly_cooling_load(
+            np.minimum(peak_cool_load, building_load.hourly_cooling_load
+            if isinstance(borefield.load, HourlyBuildingLoad) else building_load.hourly_cooling_load_simulation_period))
+        borefield.load.set_hourly_heating_load(
+            np.minimum(peak_heat_load, building_load.hourly_heating_load
+            if isinstance(borefield.load, HourlyBuildingLoad) else building_load.hourly_heating_load_simulation_period))
+        borefield.load.set_hourly_dhw_load(
+            np.minimum(peak_dhw_load, building_load.hourly_dhw_load
+            if isinstance(borefield.load, HourlyBuildingLoad) else building_load.hourly_dhw_load_simulation_period))
+
+        # calculate temperature profile, just for the results
+        borefield.calculate_temperatures(length=borefield.H, hourly=use_hourly_resolution)
+
+        # calculate relative imbalance
+        imbalance = borefield.load.imbalance / np.maximum(borefield.load.yearly_average_injection_load,
+                                                          borefield.load.yearly_average_extraction_load)
+
+        # deviation from minimum temperature
+        if abs(min(borefield.results.peak_extraction) - borefield.Tf_min) > temperature_threshold or \
+                (abs(imbalance) > 0.01 and imbalance < 0):
+            # check if it goes below the threshold
+            if min(borefield.results.peak_extraction) < borefield.Tf_min:
+                if (dhw_preferential and peak_heat_load > 0.1) \
+                        or (not dhw_preferential and peak_dhw_load <= 0.1) \
+                        or dhw_preferential is None:
+                    # first reduce the peak load in heating before touching the dhw load
+                    # if dhw_preferential is None, it is not optimised and kept constant
+                    peak_heat_load = max(0.1, peak_heat_load - 1 * max(1, 10 * (
+                            borefield.Tf_min - min(borefield.results.peak_extraction))))
+                else:
+                    peak_dhw_load = max(0.1, peak_dhw_load - 1 * max(1, 10 * (
+                            borefield.Tf_min - min(borefield.results.peak_extraction))))
+                heat_ok = False
+            else:
+                if abs(imbalance) > 0.01 and imbalance < 0:
+                    # remove imbalance
+                    if (dhw_preferential and peak_heat_load > 0.1) \
+                            or (not dhw_preferential and peak_dhw_load <= 0.1) \
+                            or dhw_preferential is None:
+                        # first reduce the peak load in heating before touching the dhw load
+                        # if dhw_preferential is None, it is not optimised and kept constant
+                        peak_heat_load = peak_heat_load * 0.99
+                    else:
+                        peak_dhw_load = peak_dhw_load * 0.99
+                elif abs(imbalance) > 0.01 and imbalance > 0:
+                    if (dhw_preferential and peak_heat_load != init_peak_heating) or (
+                            not dhw_preferential and 0.1 >= peak_dhw_load) or dhw_preferential is None:
+                        peak_heat_load = min(init_peak_heating, peak_heat_load * 1.01)
+                    else:
+                        peak_dhw_load = min(init_peak_dhw, peak_dhw_load * 1.01)
+                    if (peak_heat_load == init_peak_heating and peak_dhw_load == init_peak_dhw) or cool_ok:
+                        heat_ok = True
+                else:
+                    # imbalance small enough
+                    heat_ok = True
+        else:
+            heat_ok = True
+
+        # deviation from maximum temperature
+        if abs(np.max(borefield.results.peak_injection) - borefield.Tf_max) > temperature_threshold or \
+                (abs(imbalance) > 0.01 and imbalance > 0):
+            # check if it goes above the threshold
+            if np.max(borefield.results.peak_injection) > borefield.Tf_max:
+                peak_cool_load = max(0.1, peak_cool_load - 1 * max(1, 10 * (
+                        -borefield.Tf_max + np.max(borefield.results.peak_injection))))
+                cool_ok = False
+            else:
+                if abs(imbalance) > 0.01 and imbalance > 0:
+                    # remove imbalance
+                    peak_cool_load = peak_cool_load * 0.99
+                elif abs(imbalance) > 0.01 and imbalance < 0:
+                    peak_cool_load = min(init_peak_cooling, peak_cool_load * 1.01)
+                    if peak_cool_load == init_peak_cooling or heat_ok:
+                        cool_ok = True
+                else:
+                    # imbalance is small enough
+                    cool_ok = True
+        else:
+            cool_ok = True
+
+    # calculate external load
+    external_load = HourlyBuildingLoad(simulation_period=building_load.simulation_period)
+    external_load.set_hourly_heating_load(
+        np.maximum(0, building_load.hourly_heating_load - borefield.load.hourly_heating_load))
+    external_load.set_hourly_cooling_load(
+        np.maximum(0, building_load.hourly_cooling_load - borefield.load.hourly_cooling_load))
+    external_load.set_hourly_dhw_load(
+        np.maximum(0, building_load.hourly_dhw_load - borefield.load.hourly_dhw_load))
+
+    return borefield.load, external_load